ConfigService/requirements.txt
PyYAML
pymongo
jsonschema
python-multipart